rotate the log files from python
i was going to let logrotate deal with it, so we can also `gzip` the
files. but it turns out the files are pretty small, so gzip is not
really necessary. but i still want to rotate the files so that a power
loss doesn't kill everything we've gotten so far. so going back to the
python version